Fully functional redux and non redux tree.
