+++ demo-app/src/components/Node.tsx
@@ -1,97 +1,9 @@
 import * as React from 'react';
-import { CheckboxButton, CheckboxButtonOnChange } from './CheckboxButton';
-import { ExpandButton, ExpandButtonOnChange } from './ExpandButton';
-import { ConnectedNode } from '../internal';
-
-/**
- * Interface for the node's state property.
- */
-export interface NodeState {
-    checked?: boolean;
-    disabled?: boolean;
-    expanded?: boolean;
-    selected?: boolean;
-}
-
-export interface SelectOnChange {
-    (nodeId: string, selected: boolean): void;
-}
-
-export interface OnLazyLoad {
-    (nodeId: string): void;
-}
-
-export interface TreeData {
-    [key: string]: NodeProps;
-}
-
-/**
- * Interface for all data required from the tree root.
- */
-export interface ParentData {
-    // Non redux
-    tree: TreeData;
-
-    // Callbacks
-    checkboxOnChange: CheckboxButtonOnChange;
-    expandOnChange: ExpandButtonOnChange;
-    selectOnChange: SelectOnChange;
-    onLazyLoad: OnLazyLoad;
-    showCheckbox: boolean;
-    initSelectedNode: (nodeId: string) => void;
-
-    // Icons
-    showIcon: boolean;                 // < Determines if the icons are showed in nodes.
-    showImage: boolean;                // < Determines if images are preferred to the icons.
-    nodeIcon: string;                  // < Default icon for nodes without it.
-    checkedIcon: string;               // < The checkbox-checked icon.
-    uncheckedIcon: string;             // < The checkbox-unchecked icon.
-    partiallyCheckedIcon: string;      // < The checkbox-partially selected icon.
-    collapseIcon: string;              // < The icon for collapsing parents.
-    expandIcon: string;                // < The icon for expanding parents.
-    loadingIcon: string;               // < The loading icon when loading data with ajax.
-    errorIcon: string;                 // < The icon displayed when lazyLoading went wrong.
-    selectedIcon: string;              // < The icon for selected nodes.
-
-    // Styling
-    changedCheckboxClass: string;      // < Extra class for the changed checkbox nodes.
-    selectedClass: string;             // < Extra class for the selected nodes.
-
-    // Other
-    checkboxFirst: boolean;            // < Determines the order of the icon and the checkbox.
-    isRedux: boolean;                // < Determines if it should use the redux or the non redux version.
-}
-
-/**
- * Node properties interface.
- */
-export interface NodeProps {
-    nodeId?: string;
-    text?: string;
-    nodes?: string[];
-    state?: NodeState;
-
-    checkable?: boolean;
-    hideCheckbox?: boolean;
-
-    selectable?: boolean;
-    selectedIcon?: string;
-
-    lazyLoad?: boolean;
-    loading?: boolean; // Null when error occurred
-
-    attr?: {[key: string]: string};
-
-    // Styling
-    icon?: string;
-    iconColor?: string;
-    iconBackground?: string;
-    image?: string;
-    classes?: string;
-
-    // Private
-    parentData?: ParentData;
-}
+import {
+    ConnectedNode,
+    CheckboxButton,
+    ExpandButton, NodeProps, ParentData,
+} from '../internal';
 
 /**
  * @class Node
