+++ demo-app/src/index.tsx
@@ -13,7 +13,7 @@ ReactDOM.render(
 );
 registerServiceWorker();
 
-// Non Redux Caller
+// // Non Redux Caller
 // import * as React from 'react';
 // import * as ReactDOM from 'react-dom';
 // import App from './App-NonRedux';
